move navigation files into own folder in controllers and readers folder
