+++ lib/src/controllers/controllers.dart
@@ -11,5 +11,5 @@ import '../utils/xml_utils.dart';
 
 part 'epub_package_controller.dart';
 part 'epub_container_controller.dart';
-part 'epub_navigation_controller.dart';
-part 'epub_navigation_map_controller.dart';
+part 'navigation/epub_navigation_controller.dart';
+part 'navigation/epub_navigation_map_controller.dart';
