+++ lib/src/readers/navigation/epub_navigation_map_reader.dart
