+++ lib/src/readers/navigation/epub_navigation_reader.dart
